+++ android/app/src/main/java/feature/budget/BudgetState.kt
@@ -5,20 +5,20 @@ package com.ict.spentopia.feature.budget
 data class BudgetState(
 
     // 월 수입
-    val monthlyIncome: Int = 0,
+    val monthlyIncome: Long = 0L,
 
     // 저축 목표
-    val savingGoal: Int = 0,
+    val savingGoal: Long = 0L,
 
     // 식비 예산
-    val foodBudget: Int = 0,
+    val foodBudget: Long = 0L,
 
     // 교통 예산
-    val transportBudget: Int = 0,
+    val transportBudget: Long = 0L,
 
     // 생활 예산
-    val livingBudget: Int = 0,
+    val livingBudget: Long = 0L,
 
     // 취미 예산
-    val hobbyBudget: Int = 0
+    val hobbyBudget: Long = 0L
 )
